Class couple_iterator_t was moved into separate file
