+++ src/upload_p.hpp
@@ -22,6 +22,7 @@
 
 #include "upload.hpp"
 #include "writer.hpp"
+#include "couple_iterator.hpp"
 #include "buffered_writer.hpp"
 #include "deferred_function.hpp"
 
@@ -32,64 +33,6 @@
 #include <stdexcept>
 
 namespace elliptics {
-
-class couple_iterator_t {
-public:
-	static
-	mastermind::couple_info_t
-	create_couple_info(mastermind::groups_t groups) {
-		mastermind::couple_info_t result;
-		result.id = *std::min_element(groups.begin(), groups.end());
-		result.groups = std::move(groups);
-		return result;
-	}
-
-	couple_iterator_t(mastermind::couple_sequence_t &couple_sequence_)
-		: couple_sequence(std::move(couple_sequence_))
-		, iter(couple_sequence.begin())
-	{
-	}
-
-	couple_iterator_t(mastermind::couple_info_t single_couple_info_)
-		: iter(couple_sequence.end())
-		, single_couple_info(std::move(single_couple_info_))
-	{
-	}
-
-	couple_iterator_t(mastermind::groups_t groups)
-		: iter(couple_sequence.end())
-		, single_couple_info(create_couple_info(std::move(groups)))
-	{
-	}
-
-	bool
-	has_next() const {
-		return iter != couple_sequence.end() || single_couple_info.groups.size();
-	}
-
-	mastermind::couple_info_t
-	next() {
-		if (!has_next()) {
-			throw std::runtime_error("null couple iterator");
-		}
-
-		if (single_couple_info.groups.size()) {
-			auto result = std::move(single_couple_info);
-			single_couple_info.groups.clear();
-			return result;
-		}
-
-		auto result = *iter++;
-		return result;
-	}
-
-private:
-	mastermind::couple_sequence_t couple_sequence;
-	mastermind::couple_sequence_t::const_iterator iter;
-
-	mastermind::couple_info_t single_couple_info;
-
-};
 
 struct upload_simple_t
 	: public ioremap::thevoid::buffered_request_stream<proxy>
